move app.rb and line-bot.rb to /lib
