+++ config.ru
@@ -1,2 +1,5 @@
-require './app'
+# frozen_string_literal: true
+$LOAD_PATH.push File.expand_path("../lib", __FILE__)
+
+require 'app'
 run App
